Default the output format to circom variables
Prior this commit, the counterexample output format is defaulted to r1cs
signals (although the help page incorrectly says otherwise).
This commit fixes the help page and at the same time switches
the default format to circom variables, since the format is more
readable for clients. The flag `--raw-output` can still be used to view
the r1cs signals, which could be useful for debugging.