solvers: close ports properly (#9)
On benchmarks that are not quickly solved, I keep encountering the
error:

  subprocess: process creation failed
    system error: Too many open files; errno=24

This is because ports are not properly closed when the corresponding
process is killed. This commit fixes the problem.

In the long run, it would be better to restructure these modules to
share common code, so that the fix can be done at a single place.