fix: reimplement linear lemma to make it more efficient
Prior this commit, linear lemma depends on the CDMAP and RCDMAP
procedures.

Let's say that constraint i involves variables Xi and Yi, where Xi variables
are the linear part, and Yi variables are the non-linear parts.
That is to say, the input follows this format:

  Constraint 1: X11 X12 ... Y11 Y12 ...
  Constraint 2: X21 X22 ... Y21 Y22 ...
  ...
  Constraint n: Xn1 Xn2 ... Yn1 Yn2 ...

which is of size O(sum |Xi| + sum |Yi|).

The CDMAP procedure produces a map from a variable to a list of list of
variables. Meaning: a key can be uniquely determined if one of the list
of variables are all uniquely determined.

Thus, CDMAP produces a data of size O(sum { |Xi| (|Xi| + |Yi|) }) as the
output (where values of the same key are grouped together)

  X11: X12 X13 ... Y11 Y12 ...
  X12: X11 X13 ... Y11 Y12 ...
  ...
  X1k: X11 X12 ... Y11 Y12 ...

  X21: X22 X23 ... Y21 Y22 ...
  X22: X21 X23 ... Y21 Y22 ...
  ...

  .....

  Xn1: Xn2 Xn3 ... Yn1 Yn2 ...
  Xn2: Xn1 Xn3 ... Yn1 Yn2 ...
  ...
  Xnk: Xn1 Xn2 ... Yn1 Yn2 ...

Often, some clauses might have O(n) variables, making the size
O(n^3) if the table is dense, or O(n^2) if the table is sparse.

The RCDMAP procedure essentially consumes a map that CDMAP procedure
outputs, and produces the inverse of that map, with duplicate keys of
the input map grouped together. The size is thus the same.

RCDMAP is also used in the counter selector. E.g., given:

  1 2 3 | 4 6
  1 3 5 | 4 6

Then CDMAP produces:

  1: [2 3 4 6, 3 5 4 6]
  2: [1 3 4 6]
  3: [1 2 4 6, 1 5 4 6]
  5: [1 3 4 6]

And RCDMAP produces:

  1 3 4 6: [2, 5]
  2 3 4 6: [1]
  3 5 4 6: [1]
  1 2 4 6: [3]
  1 5 4 6: [3]

The counter selector counts that 5 occurs 2 times in the keys,
so it is weighted 2. 6, by contrast, occurs 5 times, so it is weighted 5.

-------------------------------------------------------------------

This commit aims to improve the performance by avoiding blowing up the
data size beyond linear. It also aims to preserve the existing behavior
as much as possible.

This is done by removing the CDMAP and RCDMAP procedures,
and instead employing the straightforward encoding:

  X11 X12 ... | Y11 Y12 ...
  X21 X22 ... | Y21 Y22 ...
  ...
  Xn1 Xn2 ... | Yn1 Yn2 ...

which has the linear size.

When a variable is uniquely determined, we simply cross it off from the
above table. A variable becomes propagable when it's the only one left
in the linear part, and the non-linear part is empty.

For counter selector, we produce the counter without expanding the
encoding beyond linear size. This can be done by simple multiplication:
in (Xi, Yi), each variable in Xi contributes to other variables in Xi
exactly |Xi|-1 times (not counting itself) and each variable in Yi
contributes to other variables in Xi exactly |Xi| times.

The results are not the same, however. Consider the earlier concrete example:

Our encoding:

  1 2 3 | 4 6
  1 3 5 | 4 6

RCDMAP:
  1 3 4 6: [2, 5]
  2 3 4 6: [1]
  3 5 4 6: [1]
  1 2 4 6: [3]
  1 5 4 6: [3]

The weights of 5 in both our encoding and RCDMAP agree: 2.
However, the weights of 6 do not agree.

  Our encoding produces weight 6.
  RCDMAP produces weight 5.

Arguably, our weight is more "correct", because it takes the value
part into account. I.e., in "1 3 4 6: [2, 5]", 6 contributes twice
(2 once and 5 once), so it should be counted twice.

Separately, I also believe that the counter selector is not ideal,
and plan to revamp the strategy soon, so the discrepancies will not
matter anyway.