fix: improve performance of basis2-lemma
Previously, we use `(list->set (set->list ks))` to copy the set `ks`,
but this is in fact not needed at all. `ks`, which is an immutable set,
can't be structurally mutated (although its binding can be mutated).

The copying is especially harmful when it's done in a hot loop
(the `update` function in particular), since known and unknown set could
be very large.

This change makes the large benchmark not stuck at basis2-lemma