Use temporary directory more effectively (#34)
* put smt2 files in temp dir

Now that we have a proper per-run temp directory, we can put
generated smt2 files in the temp directory, so that the files do not
clutter the main temp directory, and so that the files will be
cleaned up when Picus finishes execution.

* feat: add --noclean to prevent cleaning

Picus initially never clean up temporary files. However, the last commit
changed this so that temporary files are cleaned up. Sometimes, though,
we really do not want to clean up files (e.g. in order to generate
smt2 files for debugging purposes). This commit adds the flag to
control whether to run the clean up.