add diffing for other
In the weak correctness (default mode), output signals must be
different, so diffing the output signals are informative enough.

But in the strong corectness (when --strong is given), output signals
could be identical. It would be more helpful to diff the internal
signals to see what are different.

This commit adds diffing for the internal variables.